Update origin.md (#1111)
* Update origin.md

It now does not lie and say it returns an int, when it doesn't. It returns a string! After merge, will have @jlchmura update his references for the LPC Language server to stop me from getting angry messages from it telling me how an int and a string aren't meant to be together. Pssh.

* adding 'a'